Add `Scratch` stdlib
This implements functionality and tests for a new `Scratch` subsystem,
providing an abstraction for a mutable datastore that can be explicitly
lifecycled to an owning package, or shared among multiple packages.

Note that garbage collecting these scratch spaces is relegated to
`Pkg.gc()`, included in the next commit that bumps Pkg version.